+++ Api/Data/HokodoEntityInterface.php
@@ -12,7 +12,7 @@ namespace Hokodo\BNPL\Api\Data;
 
 use Hokodo\BNPL\Api\Data\Company\CreditLimitInterface;
 
-interface HokodoEntityInterface
+interface HokodoEntityInterface extends \Magento\Framework\Api\ExtensibleDataInterface
 {
     /**
      * Company Id getter.