+++ Model/Data/HokodoCustomer.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 namespace Hokodo\BNPL\Model\Data;
 
 use Hokodo\BNPL\Api\Data\Company\CreditLimitInterface;
+use Hokodo\BNPL\Api\Data\HokodoCustomerExtensionInterface;
 use Hokodo\BNPL\Api\Data\HokodoCustomerInterface;
 use Magento\Framework\DataObject;
 
@@ -115,5 +116,21 @@ class HokodoCustomer extends DataObject implements HokodoCustomerInterface
     {
         $this->setData(self::CREDIT_LIMIT, $credit);
         return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExtensionAttributes(): HokodoCustomerExtensionInterface
+    {
+        return $this->getData(self::EXTENSION_ATTRIBUTES_KEY);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setExtensionAttributes($extensionAttributes): self
+    {
+        return $this->setData(self::EXTENSION_ATTRIBUTES_KEY, $extensionAttributes);
     }
 }